requirements.txt

Flask==3.0.3
Flask-Cors==5.0.0
Flask-Session==0.8.0
flask-restful==0.3.10

gunicorn==20.1.0
waitress==2.1.2 # Gunicorn for Windows

python-dotenv==1.0.1
MarkupSafe==2.1.2 # Required by Azure
requests==2.24.0
identity==0.9.1 # Required for MSAL 1.24 + 

openai==1.69.0
tenacity==9.0.0

azure-identity==1.19.0
azure-storage-blob==12.23.1
azure-data-tables==12.5.0
azure-ai-translation-text==1.0.1